+++ src/main/kotlin/com/group/goyaapp/dto/request/map/MapEnterRequest.kt
@@ -1,6 +1,6 @@
 package com.group.goyaapp.dto.request.map
 
 data class MapEnterRequest(
-    val userUid: Int,
+    val userUid: Long,
     val mapId: String,
 )
