+++ src/main/kotlin/com/group/goyaapp/dto/request/quest/QuestAcceptRequest.kt
@@ -1,5 +1,5 @@
 package com.group.goyaapp.dto.request.quest
 
 data class QuestAcceptRequest(
-	val userUid: Int, val questId: String
+	val userUid: Long, val questId: String
 )
