+++ src/main/kotlin/com/group/goyaapp/dto/request/quest/QuestClearRequest.kt
@@ -1,5 +1,5 @@
 package com.group.goyaapp.dto.request.quest
 
 data class QuestClearRequest(
-	val userUid: Int, val questId: String
+	val userUid: Long, val questId: String
 )
