+++ src/main/kotlin/com/group/goyaapp/dto/request/quest/QuestLoadRequest.kt
@@ -1,5 +1,5 @@
 package com.group.goyaapp.dto.request.quest
 
 data class QuestLoadRequest(
-	val user_uid: Int
+	val user_uid: Long
 )
