+++ src/main/kotlin/com/group/goyaapp/dto/request/user/UserUpdateRequest.kt
@@ -1,5 +1,5 @@
 package com.group.goyaapp.dto.request.user
 
 data class UserUpdateRequest(
-  val userUid: Int, val nickname: String
+  val userUid: Long, val nickname: String
 )
