+++ src/main/kotlin/com/group/goyaapp/repository/QuestRepository.kt
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository
 
 interface QuestRepository : JpaRepository<Quest, Long> {
 	
-	fun findByUserUidAndQuestId(userUid: Int, questId: String): Quest?
-	fun findByUserUid(userUid: Int): List<Quest>?
+	fun findByUserUidAndQuestId(userUid: Long, questId: String): Quest?
+	fun findByUserUid(userUid: Long): List<Quest>?
 	
 }
